update game_agent.py
Tried to optimizine alpha beta as it seems to be pretty sloppy.... Tried to use an older implementation and now everything is a mess. Right now all my code sucks. Not a single innovative or clever thing and I've only gotten this far by being super literal with the transcrptions of the pseudocode.